Updated app.rputes and delted duplicates
