+++ server/src/aiProjectAdmin/aiProjectTrainModel.ts
@@ -1,5 +1,4 @@
-import type { RoutePayload } from '#types'
-import type { RouteApi, AIProjectTrainModelRequest } from '#types'
+import type { RouteApi, AIProjectTrainModelRequest, RoutePayload } from '#types'
 import { TileServerSessionsHandler } from '#src/wsisessions/TileServerSessionsHandler.ts'
 import SessionManager from '#src/wsisessions/SessionManager.ts'
 import { getDbConnection } from '#src/aiHistoDBConnection.js'
