+++ server/src/routes/saveWSIAnnotation.ts
@@ -1,5 +1,4 @@
-import type { RoutePayload, RouteApi } from '#types'
-import type { Mds3, TileSelection, SaveWSIAnnotationRequest, AIProjectAuthInfo } from '#types'
+import type { Mds3, TileSelection, SaveWSIAnnotationRequest, AIProjectAuthInfo, RoutePayload, RouteApi } from '#types'
 import { FlagStatus, SelectionPrefixes, checkSelectionType } from '#types'
 import { getDbConnection } from '#src/aiHistoDBConnection.ts'
 import type Database from 'better-sqlite3'
