+++ server/src/routes/wsimages.ts
@@ -1,10 +1,9 @@
-import type { RoutePayload, RouteApi } from '#types'
 import ky from 'ky'
 import qs from 'qs'
 import path from 'path'
 import { CookieJar } from 'tough-cookie'
 import { promisify } from 'util'
-import type { PredictionOverlay, WSImagesRequest, WSImagesResponse } from '#types'
+import type { PredictionOverlay, WSImagesRequest, WSImagesResponse, RoutePayload, RouteApi } from '#types'
 import SessionManager from '../wsisessions/SessionManager.ts'
 import type { SessionData } from '../wsisessions/SessionManager.ts'
 import { ShardManager } from '#src/sharding/ShardManager.ts'
